+++ R/predict.R
@@ -365,15 +365,30 @@ predict.stan_nma <- function(object, ...,
     # Produce nma_summary
     if (summary) {
       pred_summary <- summary_mcmc_array(pred_array, probs)
+
+      if (object$likelihood == "ordered") {
+        pred_summary <- tibble::add_column(pred_summary,
+                                           .trt = rep(preddat$.trt, each = n_cc),
+                                           .category = rep(l_cc, times = nrow(preddat)),
+                                           .before = 1)
+      } else {
+        pred_summary <- tibble::add_column(pred_summary,
+                                           .trt = preddat$.trt,
+                                           .before = 1)
+      }
+
       if (is.null(baseline)) {
         if (object$likelihood == "ordered") {
           pred_summary <- tibble::add_column(pred_summary,
                                              .study = rep(preddat$.study, each = n_cc),
                                              .before = 1)
         } else {
-          pred_summary <- tibble::add_column(pred_summary, .study = preddat$.study, .before = 1)
+          pred_summary <- tibble::add_column(pred_summary,
+                                             .study = preddat$.study,
+                                             .before = 1)
         }
       }
+
       out <- list(summary = pred_summary, sims = pred_array)
     } else {
       out <- list(sims = pred_array)
@@ -839,9 +854,14 @@ predict.stan_nma <- function(object, ...,
       if (object$likelihood == "ordered") {
         pred_summary <- tibble::add_column(pred_summary,
                                            .study = rep(preddat$.study, each = n_cc),
+                                           .trt = rep(preddat$.trt, each = n_cc),
+                                           .category = rep(l_cc, times = nrow(preddat)),
                                            .before = 1)
       } else {
-        pred_summary <- tibble::add_column(pred_summary, .study = preddat$.study, .before = 1)
+        pred_summary <- tibble::add_column(pred_summary,
+                                           .study = preddat$.study,
+                                           .trt = preddat$.trt,
+                                           .before = 1)
       }
       out <- list(summary = pred_summary, sims = pred_array)
     } else {